+++ db/schema.rb
@@ -374,8 +374,8 @@ ActiveRecord::Schema.define(version: 2022_02_16_221450) do
   end
 
   create_table "confidences", id: :serial, force: :cascade do |t|
-    t.string "confidence_object_type", null: false
     t.integer "confidence_object_id", null: false
+    t.string "confidence_object_type", null: false
     t.integer "position", null: false
     t.integer "created_by_id", null: false
     t.integer "updated_by_id", null: false
@@ -504,7 +504,7 @@ ActiveRecord::Schema.define(version: 2022_02_16_221450) do
     t.integer "encoded_dataset_record_type", null: false
     t.integer "project_id", null: false
     t.integer "import_dataset_id", null: false
-    t.integer "dataset_record_id", null: false
+    t.bigint "dataset_record_id", null: false
     t.index "import_dataset_id, encoded_dataset_record_type, \"position\", substr((value)::text, 1, 1000), dataset_record_id", name: "index_dataset_record_fields_for_filters", unique: true
     t.index ["dataset_record_id", "position"], name: "index_dataset_record_fields_on_dataset_record_id_and_position", unique: true
   end
@@ -517,8 +517,8 @@ ActiveRecord::Schema.define(version: 2022_02_16_221450) do
     t.datetime "updated_at", precision: 6, null: false
     t.integer "created_by_id", null: false
     t.integer "updated_by_id", null: false
-    t.integer "project_id"
-    t.integer "import_dataset_id"
+    t.bigint "project_id"
+    t.bigint "import_dataset_id"
     t.index ["created_by_id"], name: "index_dataset_records_on_created_by_id"
     t.index ["import_dataset_id", "type", "id"], name: "index_dataset_records_on_import_dataset_id_and_type_and_id"
     t.index ["import_dataset_id"], name: "index_dataset_records_on_import_dataset_id"
@@ -607,8 +607,8 @@ ActiveRecord::Schema.define(version: 2022_02_16_221450) do
   end
 
   create_table "documentation", id: :serial, force: :cascade do |t|
-    t.string "documentation_object_type", null: false
     t.integer "documentation_object_id", null: false
+    t.string "documentation_object_type", null: false
     t.integer "document_id", null: false
     t.integer "project_id", null: false
     t.integer "created_by_id", null: false
@@ -626,7 +626,7 @@ ActiveRecord::Schema.define(version: 2022_02_16_221450) do
   create_table "documents", id: :serial, force: :cascade do |t|
     t.string "document_file_file_name", null: false
     t.string "document_file_content_type", null: false
-    t.bigint "document_file_file_size", null: false
+    t.integer "document_file_file_size", null: false
     t.datetime "document_file_updated_at", null: false
     t.integer "project_id", null: false
     t.integer "created_by_id", null: false
@@ -835,8 +835,8 @@ ActiveRecord::Schema.define(version: 2022_02_16_221450) do
     t.string "vernacularName"
     t.string "waterBody"
     t.string "year"
-    t.string "dwc_occurrence_object_type"
     t.integer "dwc_occurrence_object_id"
+    t.string "dwc_occurrence_object_type"
     t.integer "created_by_id", null: false
     t.integer "updated_by_id", null: false
     t.integer "project_id"
@@ -1032,7 +1032,7 @@ ActiveRecord::Schema.define(version: 2022_02_16_221450) do
     t.datetime "updated_at", null: false
     t.string "image_file_file_name"
     t.string "image_file_content_type"
-    t.bigint "image_file_file_size"
+    t.integer "image_file_file_size"
     t.datetime "image_file_updated_at"
     t.integer "updated_by_id", null: false
     t.text "image_file_meta"
@@ -1052,7 +1052,7 @@ ActiveRecord::Schema.define(version: 2022_02_16_221450) do
     t.datetime "updated_at", precision: 6, null: false
     t.integer "created_by_id", null: false
     t.integer "updated_by_id", null: false
-    t.integer "project_id"
+    t.bigint "project_id"
     t.string "source_file_name"
     t.string "source_content_type"
     t.bigint "source_file_size"
@@ -1085,7 +1085,7 @@ ActiveRecord::Schema.define(version: 2022_02_16_221450) do
     t.datetime "updated_at", null: false
     t.string "type"
     t.index ["created_by_id"], name: "labels_created_by_id_index"
-    t.index ["label_object_type", "label_object_id"], name: "index_labels_on_label_object"
+    t.index ["label_object_type", "label_object_id"], name: "index_labels_on_label_object_type_and_label_object_id"
     t.index ["project_id"], name: "index_labels_on_project_id"
     t.index ["updated_by_id"], name: "labels_updated_by_id_index"
   end
@@ -1113,8 +1113,8 @@ ActiveRecord::Schema.define(version: 2022_02_16_221450) do
     t.integer "project_id", null: false
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
-    t.string "loan_item_object_type"
     t.integer "loan_item_object_id"
+    t.string "loan_item_object_type"
     t.integer "total"
     t.string "disposition"
     t.index ["created_by_id"], name: "index_loan_items_on_created_by_id"
@@ -1143,7 +1143,7 @@ ActiveRecord::Schema.define(version: 2022_02_16_221450) do
     t.datetime "updated_at", null: false
     t.string "recipient_honorific"
     t.string "recipient_country"
-    t.text "lender_address", null: false
+    t.text "lender_address", default: "Lender's address not provided.", null: false
     t.boolean "is_gift"
     t.index ["created_by_id"], name: "index_loans_on_created_by_id"
     t.index ["project_id"], name: "index_loans_on_project_id"
@@ -1351,10 +1351,10 @@ ActiveRecord::Schema.define(version: 2022_02_16_221450) do
   end
 
   create_table "origin_relationships", id: :serial, force: :cascade do |t|
-    t.string "old_object_type", null: false
     t.integer "old_object_id", null: false
-    t.string "new_object_type", null: false
+    t.string "old_object_type", null: false
     t.integer "new_object_id", null: false
+    t.string "new_object_type", null: false
     t.integer "position"
     t.integer "created_by_id", null: false
     t.integer "updated_by_id", null: false
@@ -1362,8 +1362,8 @@ ActiveRecord::Schema.define(version: 2022_02_16_221450) do
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
     t.index ["created_by_id"], name: "index_origin_relationships_on_created_by_id"
-    t.index ["new_object_type", "new_object_id"], name: "index_origin_relationships_on_new_object"
-    t.index ["old_object_type", "old_object_id"], name: "index_origin_relationships_on_old_object"
+    t.index ["new_object_type", "new_object_id"], name: "index_origin_relationships_on_new_object_type_and_new_object_id"
+    t.index ["old_object_type", "old_object_id"], name: "index_origin_relationships_on_old_object_type_and_old_object_id"
     t.index ["project_id"], name: "index_origin_relationships_on_project_id"
     t.index ["updated_by_id"], name: "index_origin_relationships_on_updated_by_id"
   end
@@ -1438,8 +1438,8 @@ ActiveRecord::Schema.define(version: 2022_02_16_221450) do
   end
 
   create_table "pinboard_items", id: :serial, force: :cascade do |t|
-    t.string "pinned_object_type", null: false
     t.integer "pinned_object_id", null: false
+    t.string "pinned_object_type", null: false
     t.integer "user_id", null: false
     t.integer "project_id", null: false
     t.integer "position", null: false
@@ -1451,7 +1451,7 @@ ActiveRecord::Schema.define(version: 2022_02_16_221450) do
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
     t.index ["created_by_id"], name: "index_pinboard_items_on_created_by_id"
-    t.index ["pinned_object_type", "pinned_object_id"], name: "index_pinboard_items_on_pinned_object"
+    t.index ["pinned_object_type", "pinned_object_id"], name: "index_pinboard_items_on_pinned_object_type_and_pinned_object_id"
     t.index ["position"], name: "index_pinboard_items_on_position"
     t.index ["project_id"], name: "index_pinboard_items_on_project_id"
     t.index ["updated_by_id"], name: "index_pinboard_items_on_updated_by_id"
@@ -1503,7 +1503,7 @@ ActiveRecord::Schema.define(version: 2022_02_16_221450) do
     t.datetime "updated_at", null: false
     t.integer "created_by_id", null: false
     t.integer "updated_by_id", null: false
-    t.jsonb "preferences", default: "{}", null: false
+    t.jsonb "preferences", default: {}, null: false
     t.string "api_access_token"
     t.index ["created_by_id"], name: "index_projects_on_created_by_id"
     t.index ["updated_by_id"], name: "index_projects_on_updated_by_id"
@@ -1511,8 +1511,8 @@ ActiveRecord::Schema.define(version: 2022_02_16_221450) do
 
   create_table "protocol_relationships", id: :serial, force: :cascade do |t|
     t.integer "protocol_id", null: false
-    t.string "protocol_relationship_object_type", null: false
     t.integer "protocol_relationship_object_id", null: false
+    t.string "protocol_relationship_object_type", null: false
     t.integer "position", null: false
     t.integer "created_by_id", null: false
     t.integer "updated_by_id", null: false
@@ -1681,7 +1681,7 @@ ActiveRecord::Schema.define(version: 2022_02_16_221450) do
   end
 
   create_table "sled_images", force: :cascade do |t|
-    t.integer "image_id", null: false
+    t.bigint "image_id", null: false
     t.jsonb "metadata"
     t.jsonb "object_layout"
     t.integer "cached_total_rows"
@@ -1689,7 +1689,7 @@ ActiveRecord::Schema.define(version: 2022_02_16_221450) do
     t.integer "cached_total_collection_objects", default: 0, null: false
     t.integer "created_by_id", null: false
     t.integer "updated_by_id", null: false
-    t.integer "project_id"
+    t.bigint "project_id"
     t.datetime "created_at", precision: 6, null: false
     t.datetime "updated_at", precision: 6, null: false
     t.index ["created_by_id"], name: "index_sled_images_on_created_by_id"
@@ -1769,8 +1769,8 @@ ActiveRecord::Schema.define(version: 2022_02_16_221450) do
     t.string "boundary_finder", null: false
     t.boolean "has_border", null: false
     t.string "layout", null: false
-    t.jsonb "metadata_map", default: "{}", null: false
-    t.jsonb "specimen_coordinates", default: "{}", null: false
+    t.jsonb "metadata_map", default: {}, null: false
+    t.jsonb "specimen_coordinates", default: {}, null: false
     t.integer "project_id", null: false
     t.integer "created_by_id", null: false
     t.integer "updated_by_id", null: false